separate all navbar functions, initial test on async function. needs babel-polyfill with async/await
