+++ app/javascript/react/src/components/UserSignedOut.js
@@ -1,32 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router';
-import swal from 'sweetalert';
+import { signIn } from '../POJOs/NavBarFunctions'
+
 
 const UserSignedOut = props => {
-
-  let signIn = () => {
-    swal({
-      title: "Sign In",
-      icon: "info",
-      buttons: {
-        google: {
-          text: 'Sign In with Google',
-          value: 'google'
-        },
-        github: {
-          text: 'Sign In with Github',
-          value: 'github'
-        }
-      }
-    })
-    .then(value => {
-      if(value == 'github'){
-        window.location.replace("/auth/github")
-      } else if (value == 'google'){
-        window.location.replace("/auth/google")
-      }
-    })
-  }
 
   return(
     <div className='grid-container'>
